shared/netutils: Factor mp_hal_get_mac_ascii out of ports.
Move the identical helper from stm32/rp2/mimxrt/renesas-ra/alif into
shared/netutils and assert the hex character slice stays within the
12-character MAC representation.

Fixes #19668.
