+++ ports/renesas-ra/Makefile
@@ -154,6 +154,7 @@ MPY_CROSS_FLAGS += -march=armv7m
 SHARED_SRC_C += $(addprefix shared/,\
 	libc/string0.c \
 	netutils/dhcpserver.c \
+	netutils/get_mac_ascii.c \
 	netutils/netutils.c \
 	netutils/trace.c \
 	readline/readline.c \
